nastavena pristupova prava

+++ src/OpenIntranet/plugins/attendance.py
@@ -10,11 +10,14 @@ from plugins import BaseHandler, get_dpp_params, BaseHandlerOwnCloud
 from plugins.helpers import database_attendance as adb, report_generation, database_reports as rdb, owncloud_utils
 from plugins.helpers import database_user as udb
 from plugins.helpers import str_ops
-from plugins.helpers.exceptions import BadInputHTTPError, MissingInfoHTTPError
+from plugins.helpers.exceptions import BadInputHTTPError, MissingInfoHTTPError, ForbiddenHTTPError
 from plugins.helpers.finance import calculate_tax
 from plugins.helpers.math_utils import floor_to_half
 from plugins.helpers.mdoc_ops import get_user_days_of_vacation_in_year
 from plugins.helpers.owncloud_utils import generate_reports_directory_path, get_file_url
+
+ROLE_SUDO = "users-sudo"
+ROLE_ACCOUNTANT = "users-accountant"
 
 
 def make_handlers(plugin_name, plugin_namespace):
@@ -51,13 +54,12 @@ def plug_info():
     }
 
 
-# TODO zkontrolovat práva
 class HomeHandler(BaseHandler):
 
     def get(self):
         current_user_id = self.actual_user["_id"]
 
-        if self.is_authorized(['users-editor', 'sudo-users']):
+        if self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT]):
             self.render('attendance.home-sudo.hbs')
         else:
             self.redirect(f"/attendance/u/{current_user_id}")
@@ -236,8 +238,8 @@ class AttendanceCalculator:
         return self.year_max_hours - self.year_hours_worked
 
 
-# TODO doplnit práva
 class ApiAdminMonthTableHandler(BaseHandler):
+    role_module = [ROLE_SUDO, ROLE_ACCOUNTANT]
 
     def get(self, date):
         date = str_ops.datetime_from_iso_str(date)
@@ -267,8 +269,8 @@ class ApiAdminMonthTableHandler(BaseHandler):
         self.write(bson.json_util.dumps(rows))
 
 
-# TODO doplnit práva
 class ApiAdminYearTableHandler(BaseHandler):
+    role_module = [ROLE_SUDO, ROLE_ACCOUNTANT]
 
     def get(self, date):
         date = str_ops.datetime_from_iso_str(date)
@@ -292,8 +294,8 @@ class ApiAdminYearTableHandler(BaseHandler):
         self.write(bson.json_util.dumps(rows))
 
 
-# TODO doplnit práva
 class ApiAdminReportsTableHandler(BaseHandler):
+    role_module = [ROLE_SUDO, ROLE_ACCOUNTANT]
 
     def get(self, date):
         date = str_ops.datetime_from_iso_str(date).replace(month=1, day=1, hour=0, minute=0, second=0, microsecond=0)
@@ -320,11 +322,13 @@ class ApiAdminReportsTableHandler(BaseHandler):
         self.write(bson.json_util.dumps(rows))
 
 
-# TODO doplnit práva
 class UserAttendancePageHandler(BaseHandler):
 
     def get(self, user_id, date_str=None):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="zobrazení docházky jiného uživatele")
 
         date = str_ops.datetime_from_iso_str(date_str)
         if not date:
@@ -370,12 +374,14 @@ class UserAttendancePageHandler(BaseHandler):
         self.render("attendance.home.hbs", **template_params)
 
 
-# TODO doplnit práva
 class ApiCalendarHandler(BaseHandler):
 
-    # TODO nemělo by toto být get?
-    def post(self, user_id, date):
+    def get(self, user_id, date):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="zobrazení kalendáře jiného uživatele")
+
         # TODO je potřeba v kalendáři mít přístup i k assignmentům
         month = str_ops.datetime_from_iso_str(date).replace(day=1)
         num_days_in_month = calendar.monthrange(month.year, month.month)[1]
@@ -414,9 +420,11 @@ class ApiCalendarHandler(BaseHandler):
 
 class ApiMonthInfoHandler(BaseHandler):
 
-    # TODO nemělo by toto být get?
-    def post(self, user_id, date):
+    def get(self, user_id, date):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="zobrazení informací o měsíci jiného uživatele")
 
         month = str_ops.datetime_from_iso_str(date).replace(day=1)
 
@@ -471,12 +479,14 @@ class WorkspanBaseHandler(BaseHandler):
         return True
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiAddWorkspanHandler(WorkspanBaseHandler):
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="přidání docházky jiného uživatele")
 
         req = self.request.body.decode("utf-8")
         form_data = bson.json_util.loads(req)
@@ -502,12 +512,14 @@ class ApiAddWorkspanHandler(WorkspanBaseHandler):
         adb.add_user_workspan(self.mdb, user_id, workspan)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiEditMonthWorkspansHandler(BaseHandler):
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="editace docházky jiného uživatele")
 
         req = self.request.body.decode("utf-8")
         form_data = bson.json_util.loads(req)
@@ -533,12 +545,15 @@ class ApiEditMonthWorkspansHandler(BaseHandler):
                 adb.add_user_workspan(self.mdb, user_id, workspan)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiAddVacationHandler(BaseHandler):
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="přidání dovolené jinému uživateli")
+
         # TODO hlídat aby nové dovolené byly v budoucnosti, nelze přidat dovolenou zpětně
         req = self.request.body.decode("utf-8")
         data = bson.json_util.loads(req)
@@ -562,12 +577,14 @@ class ApiAddVacationHandler(BaseHandler):
         adb.add_user_vacation(self.mdb, user_id, vacation)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiInterruptVacationHandler(BaseHandler):
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="přerušení dovolené jiného uživatele")
 
         req = self.request.body.decode("utf-8")
         data = json.loads(req)
@@ -581,22 +598,26 @@ class ApiInterruptVacationHandler(BaseHandler):
         adb.interrupt_user_vacation(self.mdb, user_id, data["_id"], interruption_date)
 
 
-# TODO doplnit práva
 class ApiDeleteWorkspanHandler(BaseHandler):
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
 
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="odstranění docházky jiného uživatele")
+
         workspan_id = self.request.body.decode("utf-8")
         adb.delete_user_workspan(self.mdb, user_id, workspan_id)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiCloseMonthHandler(BaseHandler):
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
+
+        if not self.is_authorized([ROLE_SUDO, ROLE_ACCOUNTANT], specific_users=[user_id]):
+            raise ForbiddenHTTPError(operation="uzavření měsíce jiného uživatele")
 
         month_date_iso = self.request.body.decode("utf-8")
         month_date = str_ops.datetime_from_iso_str(month_date_iso)
@@ -610,9 +631,9 @@ class ApiCloseMonthHandler(BaseHandler):
         adb.close_month(self.mdb, user_id, month_date)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiReopenMonthHandler(BaseHandler):
+    role_module = [ROLE_SUDO, ROLE_ACCOUNTANT]
 
     def post(self, user_id):
         user_id = ObjectId(user_id)
@@ -625,9 +646,9 @@ class ApiReopenMonthHandler(BaseHandler):
         rdb.set_report_out_of_date(self.mdb, month_date, rdb.HOURS_WORKED_REPORT)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiGenerateAccountantReportHandler(BaseHandlerOwnCloud):
+    role_module = [ROLE_SUDO, ROLE_ACCOUNTANT]
 
     async def post(self):
         month_date_iso = self.request.body.decode("utf-8")
@@ -675,9 +696,9 @@ class ApiGenerateAccountantReportHandler(BaseHandlerOwnCloud):
             rdb.save_report_file_id(self.mdb, month_date, owncloud_id, rdb.ACCOUNTANT_REPORT)
 
 
-# TODO doplnit práva
 # TODO validovat vstup
 class ApiGenerateHoursWorkedReportHandler(BaseHandlerOwnCloud):
+    role_module = [ROLE_SUDO, ROLE_ACCOUNTANT]
 
     async def post(self):
         month_date_iso = self.request.body.decode("utf-8")
